tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_update.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_update</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      var cache = window.applicationCache;

      test(function() {
        try {
          cache.update()
          assert_true(true, "update method test")
        } catch (e) {
          assert_unreached("update method failed.");
        }
      });
    </script>
  </body>
</html>
